+++ ballast-core/src/commonMain/kotlin/com/copperleaf/ballast/core/FifoInputStrategy.kt
@@ -6,7 +6,6 @@ import com.copperleaf.ballast.Queued
 import kotlinx.coroutines.channels.BufferOverflow
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.collect
 
 /**
  * A sequential first-in-first-out strategy for processing inputs, suitable for background processing. As inputs will be
@@ -22,20 +21,30 @@ import kotlinx.coroutines.flow.collect
  * Since we know only 1 Input is being procced at a time, if an input gets cancelled partway through its processing, the
  * ViewModel state will roll back to prevent the ViewModel from being left in a bad state.
  */
-public class FifoInputStrategy : InputStrategy {
+public class FifoInputStrategy<Inputs : Any, Events : Any, State : Any> private constructor(): InputStrategy<Inputs, Events, State> {
 
-    override fun <T> createQueue(): Channel<T> {
+    override fun createQueue(): Channel<Queued<Inputs, Events, State>> {
         return Channel(Channel.BUFFERED, BufferOverflow.SUSPEND)
     }
 
     override val rollbackOnCancellation: Boolean = true
 
-    override suspend fun <Inputs : Any, Events : Any, State : Any> InputStrategyScope<Inputs, Events, State>.processInputs(
+    override suspend fun InputStrategyScope<Inputs, Events, State>.processInputs(
         filteredQueue: Flow<Queued<Inputs, Events, State>>,
     ) {
         filteredQueue
             .collect { queued ->
                 acceptQueued(queued, DefaultGuardian())
             }
+    }
+
+    public companion object {
+        public operator fun invoke() : FifoInputStrategy<Any, Any, Any> {
+            return FifoInputStrategy()
+        }
+
+        public fun <Inputs : Any, Events : Any, State : Any> typed(): FifoInputStrategy<Inputs, Events, State> {
+            return FifoInputStrategy()
+        }
     }
 }