+++ examples/mpp/src/commonMain/kotlin/com/copperleaf/ballast/examples/bgg/repository/BggRepositoryImpl.kt
@@ -1,12 +1,15 @@
 package com.copperleaf.ballast.examples.bgg.repository
 
 import com.copperleaf.ballast.BallastViewModelConfiguration
+import com.copperleaf.ballast.build
 import com.copperleaf.ballast.examples.bgg.api.BggApi
 import com.copperleaf.ballast.examples.bgg.models.BggHotListItem
 import com.copperleaf.ballast.examples.bgg.models.HotListType
 import com.copperleaf.ballast.repository.BallastRepository
 import com.copperleaf.ballast.repository.bus.EventBus
 import com.copperleaf.ballast.repository.cache.Cached
+import com.copperleaf.ballast.repository.withRepository
+import com.copperleaf.ballast.withViewModel
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
@@ -19,12 +22,14 @@ class BggRepositoryImpl(
 ) : BallastRepository<BggRepositoryContract.Inputs, BggRepositoryContract.State>(
     coroutineScope = coroutineScope,
     eventBus = eventBus,
-    configBuilder = configBuilder
-        .apply {
-            this.inputHandler = BggRepositoryInputHandler(eventBus, api)
-            this.initialState = BggRepositoryContract.State()
-            this.name = "Bgg Repository"
-        },
+    config = configBuilder
+        .withViewModel(
+            inputHandler = BggRepositoryInputHandler(eventBus, api),
+            initialState = BggRepositoryContract.State(),
+            name = "Bgg Repository",
+        )
+        .withRepository()
+        .build()
 ), BggRepository {
     override fun clearAllCaches() {
         trySend(BggRepositoryContract.Inputs.ClearCaches)